Moved dialogs into their own folder
