+++ src/modules/Preservation/views/ScopeOverview/ScopeOverview.tsx
@@ -25,26 +25,26 @@ import React, { useEffect, useRef, useState } from 'react';
 import { Button } from '@equinor/eds-core-react';
 import CacheService from '@procosys/core/services/CacheService';
 import { Canceler } from '@procosys/http/HttpClient';
-import CompleteDialog from './CompleteDialog';
+import CompleteDialog from './Dialogs/CompleteDialog';
 import Dropdown from '../../../../components/Dropdown';
 import EdsIcon from '../../../../components/EdsIcon';
 import Flyout from './../../../../components/Flyout';
 import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
 import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp';
 import OptionsDropdown from '../../../../components/OptionsDropdown';
-import PreservedDialog from './PreservedDialog';
+import PreservedDialog from './Dialogs/PreservedDialog';
 import { ProjectDetails } from '../../types';
 import Qs from 'qs';
-import RemoveDialog from './RemoveDialog';
-import RescheduleDialog from './RescheduleDialog';
+import RemoveDialog from './Dialogs/RemoveDialog';
+import RescheduleDialog from './Dialogs/RescheduleDialog';
 import ScopeFilter from './ScopeFilter/ScopeFilter';
 import Spinner from '@procosys/components/Spinner';
-import StartPreservationDialog from './StartPreservationDialog';
+import StartPreservationDialog from './Dialogs/StartPreservationDialog';
 import TagFlyout from './TagFlyout/TagFlyout';
 import { Tooltip } from '@material-ui/core';
-import TransferDialog from './TransferDialog';
+import TransferDialog from './Dialogs/TransferDialog';
 import { Typography } from '@equinor/eds-core-react';
-import VoidDialog from './VoidDialog';
+import VoidDialog from './Dialogs/VoidDialog';
 import { showModalDialog } from '../../../../core/services/ModalDialogService';
 import { showSnackbarNotification } from '../../../../core/services/NotificationService';
 import { tokens } from '@equinor/eds-tokens';
